+++ data-filter/lib/filter/order-rules/default.order-rule.ts
@@ -7,7 +7,7 @@ export interface DefaultOrderRuleDefinition {
 export class DefaultOrderRule {
     private _type = "default_order_rule";
 
-    constructor(public order: DefaultOrderRuleDefinition) {}
+    constructor(public order: DefaultOrderRuleDefinition | DefaultOrderRuleDefinition[]) {}
 
     public static validate(definition: DefaultOrderRuleDefinition) {
         return definition["_type"] === "default_order_rule";