+++ data-filter/lib/filter/order-rules/index.ts
@@ -1,2 +1,4 @@
 export * from "./enum.order-rule";
 export * from "./if.order-rule";
+export * from "./is-null.order-rule";
+export * from "./default.order-rule";